refactor(gui): adopt the Tabs primitive everywhere; delete SidebarTabs
The library panel, chat/console sidebars, server list, and the settings
modal's three hand-copied tab buttons all move to ui/Tabs. Every tablist
now has an accessible name and a working active underline — the old
implementations set border-none alongside border-b-2, so their active
indicator never rendered and 'active' fell back to blue text, one of the
two competing tab treatments the restyle retires.
