+++ src/components/ConsoleInfoPanel/ConsoleInfoPanel.tsx
@@ -1,7 +1,7 @@
 import { FC, useState, useEffect, useCallback, useRef } from 'react';
 import { Copy, StopCircle } from 'lucide-react';
 import { ChatPageTabId, CHAT_PAGE_TABS } from '../../pages/chatTabs';
-import SidebarTabs from '../ModelLibraryPanel/SidebarTabs';
+import { Tabs } from '../ui/Tabs';
 import { useServerState } from '../../services/serverEvents';
 import { Icon } from '../ui/Icon';
 import { Button } from '../ui/Button';
@@ -168,10 +168,11 @@ const ConsoleInfoPanel: FC<ConsoleInfoPanelProps> = ({
       <div className="p-base border-b border-border bg-background shrink-0">
         {/* View Tabs */}
         <div className="mb-md">
-          <SidebarTabs<ChatPageTabId>
+          <Tabs<ChatPageTabId>
             tabs={CHAT_PAGE_TABS}
-            activeTab={activeTab}
-            onTabChange={onTabChange}
+            activeId={activeTab}
+            onChange={onTabChange}
+            aria-label="Chat views"
           />
         </div>
