+++ src/components/ModelLibraryPanel/ModelLibraryPanel.tsx
@@ -1,7 +1,10 @@
 import { FC, useState } from 'react';
 import { BookOpenText, PlusCircle, RefreshCcw, SlidersHorizontal } from 'lucide-react';
 import { GgufModel, ServerInfo, HfModelSummary, ModelFilterOptions } from '../../types';
-import SidebarTabs, { SidebarTabId, SidebarTab } from './SidebarTabs';
+import { Tabs, type TabItem } from '../ui/Tabs';
+
+// Legacy id union kept here after SidebarTabs was replaced by ui/Tabs.
+export type SidebarTabId = 'models' | 'add';
 import ModelsListContent from './ModelsListContent';
 import AddDownloadContent, { AddDownloadSubTab } from './AddDownloadContent';
 import ProxyControl from '../ProxyControl';
@@ -46,7 +49,7 @@ interface ModelLibraryPanelProps {
   onTabChange?: (tab: SidebarTabId) => void;
 }
 
-const SIDEBAR_TABS: SidebarTab[] = [
+const SIDEBAR_TABS: TabItem<SidebarTabId>[] = [
   { id: 'models', label: 'Your Models', icon: <BookOpenText size={18} /> },
   { id: 'add', label: 'Add Models', icon: <PlusCircle size={18} /> },
 ];
@@ -104,10 +107,12 @@ const ModelLibraryPanel: FC<ModelLibraryPanelProps> = ({
     return (
       <div className="flex flex-col overflow-y-auto overflow-x-hidden border-b border-border relative flex-1 bg-surface md:h-full md:min-h-0 md:border-b-0 md:border-r">
         <div className="p-base border-b border-border bg-background shrink-0">
-          <SidebarTabs
+          <Tabs
             tabs={SIDEBAR_TABS}
-            activeTab={activeTab}
-            onTabChange={handleTabChange}
+            activeId={activeTab}
+            onChange={handleTabChange}
+            aria-label="Library sections"
+            className="mb-md"
           />
         </div>
         <div className="flex-1 min-h-0 overflow-y-auto overflow-x-hidden flex flex-col">
@@ -149,10 +154,12 @@ const ModelLibraryPanel: FC<ModelLibraryPanelProps> = ({
   return (
     <div className="flex flex-col overflow-y-auto overflow-x-hidden border-b border-border relative flex-1 bg-surface md:h-full md:min-h-0 md:border-b-0 md:border-r">
       <div className="p-base border-b border-border bg-background shrink-0">
-        <SidebarTabs
+        <Tabs
           tabs={SIDEBAR_TABS}
-          activeTab={activeTab}
-          onTabChange={handleTabChange}
+          activeId={activeTab}
+          onChange={handleTabChange}
+          aria-label="Library sections"
+          className="mb-md"
           rightContent={headerActions}
         />
